Removed categories and merged them into filters.
Most importantly, filter selection and deselection seems to be working correctly.  This is only verifiable through the debug console and by the updated counts on the filter display in the filter grid sidebar.  When a filter is selected, the product counts decrement, and when deselected they increment.

I'm ready now to actually show the product images in the grid.  Finally we are getting somewhere!